:sparkles: Handle GithubException in send_invitation route
This commit improves the `send_invitation` function in `join.py` by adding error handling for potential exceptions when trying to send GitHub invitations. Specifically, if a `GithubException` is encountered indicating that the user is already a member of the organization, it logs the event and returns an appropriate message. Other `GithubException` instances or `ValueError` will still cause the system to abort and log the error. Relevant tests have been added to ensure these exceptions are properly handled.

Please note that this commit also includes minor changes to import statements and formatting which do not affect functionality.