:sparkles: Show custom page when user already a member
This commit introduces a new error handling method when a user tries to join an organization they're already part of. Instead of returning an abort statement with a 400 status code, the update now renders a custom page (`already-a-member.html`) informing the user that they are already a member in a more user-friendly way. Test cases related to this update were also adjusted accordingly.